+++ sumo/src/microsim/output/MSFullExport.cpp
@@ -140,7 +140,7 @@ MSFullExport::writeTLS(OutputDevice& of, SUMOTime /* timestep */) {
         }
 
         std::string lane_output = "";
-        for (int i1 = 0; i1 < laneIDs.size(); ++i1) {
+        for (int i1 = 0; i1 < (int)laneIDs.size(); ++i1) {
             lane_output += laneIDs[i1] + " ";
         }
