+++ sumo/src/netbuild/NBDistrict.cpp
@@ -176,13 +176,13 @@ NBDistrict::replaceOutgoing(const EdgeVector& which, NBEdge* const by) {
 void
 NBDistrict::removeFromSinksAndSources(NBEdge* const e) {
     int i;
-    for (i = 0; i < mySinks.size(); ++i) {
+    for (i = 0; i < (int)mySinks.size(); ++i) {
         if (mySinks[i] == e) {
             mySinks.erase(mySinks.begin() + i);
             mySinkWeights.erase(mySinkWeights.begin() + i);
         }
     }
-    for (i = 0; i < mySources.size(); ++i) {
+    for (i = 0; i < (int)mySources.size(); ++i) {
         if (mySources[i] == e) {
             mySources.erase(mySources.begin() + i);
             mySourceWeights.erase(mySourceWeights.begin() + i);
